requirements.txt
# Core Dependencies
pydantic>=2.0
apscheduler
openai
google-genai
python-dotenv
httpx
streamlit==1.56.0
neonize==0.3.16.post0
fastapi
uvicorn[standard]
psutil
cryptography
pytz
psycopg2-binary
redis>=5.0.0
mcp

# Dev Dependencies
pytest
black
isort
flake8
